Refactor: leaveChatParticipant의 반환값 string ->LeaveChatParticipantResponse
