+++ src/main/java/com/capstone/smutaxi/dto/responses/ResponseFactory.java
@@ -6,6 +6,8 @@ import com.capstone.smutaxi.dto.UserDto;
 import com.capstone.smutaxi.dto.responses.auth.EmailVerificationResponse;
 import com.capstone.smutaxi.dto.responses.auth.JoinResponse;
 import com.capstone.smutaxi.dto.responses.auth.LoginResponse;
+import com.capstone.smutaxi.dto.responses.chat.ChatRoomMessageResponse;
+import com.capstone.smutaxi.dto.responses.chat.UserJoinedChatRoomResponse;
 import com.capstone.smutaxi.dto.responses.match.MatchCancelResponse;
 import com.capstone.smutaxi.dto.responses.match.MatchingResponse;
 import com.capstone.smutaxi.dto.responses.rally.RallyInformationDto;
@@ -81,7 +83,7 @@ public class ResponseFactory {
                 .build();
     }
 
-    public static ChatRoomMessageResponse createChatRoomMessageResponse(Boolean success, String message,Long chatParticipantId, List<MessageDto> messageDtoList){
+    public static ChatRoomMessageResponse createChatRoomMessageResponse(Boolean success, String message, Long chatParticipantId, List<MessageDto> messageDtoList){
         return ChatRoomMessageResponse.builder()
                 .success(success)
                 .message(message)
